Address codex review: no-activate opens don't steal focus, deterministic mid-close rejection, fail-closed nil context
- show(activateApp: false) now only orders the window in (no makeKey, no
  parent-window raise, no unhide/activate), honoring the socket
  no-focus-steal contract for settings.open --activate=false.
- SettingsHostWindow records close-begin so show() deterministically
  refuses to reuse a dying window regardless of notification-observer
  order; demolish/strip split keeps mid-close teardown re-entrant-safe.
- settings.open with a missing control-system context now returns an
  unavailable error instead of synthesizing opened=true.
- Pure geometry statics moved to SettingsWindowGeometry.swift so the
  presenter stays under the 500-line budget threshold.
